[MIPS32] Use version of rsGetAllocation() that returns a pointer
Based on change from AOSP:
https://android-review.googlesource.com/158862/

Similar to IA-32, MIPS32 has a different struct return calling convention
that uses a hidden parameter. As that becomes the first argument to the
function, all user arguments are shifted down by one.
Thus, we opt for the version of rsGetAllocation() that emulates it with a
pointer.

This fixes issues with live wallpapers (Bubbles, Black Hole, Phase Beam)
and it fixes cts/GetAllocationTest.java for MIPS32.

Change-Id: I62d326916287616bde067ba17f59f53b98b458e5