[MIPS] Workaround a bug in clang when compiling with -mmsa
While compiling this project with -mmsa we discovered a bug in clang/llvm where
it does not honor -mno-odd-spreg switch and generates a code which uses odd
single precision registers. Bug has been reported to dmz-portal:
https://dmz-portal.mips.com/bugz/show_bug.cgi?id=2442

As a workaround disable building this project with -mmsa.

Build fails with the following message:
/tmp/rsCpuIntrinsicConvolve3x3-a10162.s: Assembler messages:
/tmp/rsCpuIntrinsicConvolve3x3-a10162.s:582: Error: float register should be even, was 3
/tmp/rsCpuIntrinsicConvolve3x3-a10162.s:607: Error: float register should be even, was 5
/tmp/rsCpuIntrinsicConvolve3x3-a10162.s:652: Error: float register should be even, was 5
/tmp/rsCpuIntrinsicConvolve3x3-a10162.s:659: Error: float register should be even, was 5

Change-Id: I594a46bc00137bf86478ad621238d436f23fb332
Signed-off-by: Goran Ferenc <goran.ferenc@imgtec.com>